Switch to actual Debian installer method by using minimal net install image
Default Ubuntu Desktop installs use Ubiquity installer. Which is not flexible at
all - it uses pre-built desktop system image instead of installing all packages
from debs. And that causes a lot of headaches when trying to change/lock kernel
version and similar. Documentation is also a little misleading most of the time.